transcriber/requirements.txt
streamlit
git+https://github.com/openai/whisper.git
ffmpeg-python
python-docx
torch
reportlab
